fix: resolve critical bugs and optimize movie detail endpoint performance
  - Fix ReportBackgroundWorker: batch dispatch, graceful shutdown, DateTime.UtcNow
  - Fix decimal→int truncation in UpdateBalance (500.99 was silently becoming 500)
  - Fix _keyLocks memory leak in CacheService
  - Replace fire-and-forget cache calls with await in AuthService and RentalService
  - Replace EF Core multi-table JOIN with Dapper QueryMultiple on GET /api/Movies/{id}
    (eliminates 11-12s cold-start latency, no longer loads user password hashes)
  - Update startup warmup to cover new Dapper query path

+++ FilmKirala.Api/Controller/AuthController.cs
@@ -37,8 +37,7 @@ namespace FilmKirala.Api.Controllers
         [HttpPost("update-balance")]
         public async Task<IActionResult> AdminUpdateBalance([FromBody] UpdateBalanceDto request)
         {
-            //  decimal olan request.NewBalance'ı int'e cast ettim
-            await _authService.UpdateUserBalanceAsync(request.Email, (int)request.NewBalance);
+            await _authService.UpdateUserBalanceAsync(request.Email, request.NewBalance);
 
             return Ok(new { Message = "User balance updated successfully." });
         }
