+++ FilmKirala.Application/DTOs/UpdateBalanceDto.cs
@@ -3,6 +3,6 @@
     public class UpdateBalanceDto
     {
         public required string Email { get; set; }
-        public decimal NewBalance { get; set; }
+        public int NewBalance { get; set; }
     }
 }
